タイプ選択のモーダルを追加
